+++ example-project/index.html
@@ -6,6 +6,10 @@
   </head>
   <body>
     <button id="greenFlag">Green Flag</button>
+    <label for="stage-size">
+      <input type="range" id="stage-size" name="stage-size" min="480" max="960" value="480">
+      Stage size
+    </label>
     <div id="project" style="border: 1px solid black"></div>
 
     <script src="index.mjs" type="module"></script>
