--- renderer-test/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Example Project</title>
-  </head>
-  <body>
-    <button id="greenFlag">Green Flag</button>
-    <label for="stage-size">
-      <input type="range" id="stage-size" name="stage-size" min="480" max="960" value="480">
-      Stage size
-    </label>
-    <div id="project" style="border: 1px solid black"></div>
-
-    <script src="index.mjs" type="module"></script>
-  </body>
-</html>
